knowledge/foundry-iq/samples/kb-over-ai-search/requirements.txt
azure-search-documents==11.7.0b2
azure-identity
openai
python-dotenv
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.0